requirements.txt
Flask
gunicorn
google-generativeai
flask-cors
